Minor bugs in functions for normalization removed. ImageLoading changed for my purposes(gotta change back later). Testing bmp via CUDA Normalization returns something very wrong. Unit testing changed, but still doesn't work.

+++ Code/CUDAFingerprinting.GPU.Common/ImageLoading.cu
@@ -53,7 +53,7 @@ int* loadBmp(BMPHeader* header, char* filename)
 	return transformedPixels;
 }
 
-void saveBmp(int* data, BMPHeader* header, char* filename)
+void saveBmp(float* data, BMPHeader* header, char* filename)
 {
 	FILE *output;
